+++ packages/project-builder-lib/src/web/hooks/index.ts
@@ -1,7 +1,7 @@
-export * from './useBlockDirtyFormNavigate.js';
-export * from './useBlockerDialog.js';
-export * from './useBlockUnsavedChangesNavigate.js';
-export * from './useErrorHandler.js';
-export * from './usePluginEnhancedSchema.js';
-export * from './useProjectDefinition.js';
-export * from './useResettableForm.js';
+export * from './use-block-dirty-form-navigate.js';
+export * from './use-blocker-dialog.js';
+export * from './use-block-unsaved-changes-navigate.js';
+export * from './use-error-handler.js';
+export * from './use-plugin-enhanced-schema.js';
+export * from './use-project-definition.js';
+export * from './use-resettable-form.js';
